pyproject.toml
[tool.poetry]
name = "demeter_utils"
version = "1.0.0"
description = "Python library that provides wrapper functions for interacting with the demeter database at a higher level."
authors = []

[tool.poetry.dependencies]
python = "~3.10.4"
click = "*"
matplotlib = "^3.7.1"
psycopg2 = "^2.9.3"
pyarrow = "^14.0.1"
scipy = "^1.10.0"
seaborn = "^0.12.2"
demeter = {git = "ssh://git@github.com/SenteraLLC/demeter.git", branch = "main"}
gql_utils = { git = "ssh://git@github.com/SenteraLLC/gql-utils.git", branch = "main"}

[tool.poetry.scripts]
demeter_utils = "demeter_utils.cli:cli"

[tool.poetry.dev-dependencies]
pytest = "*"
pre_commit = "*"
sure = "^2.0.1"
flake8 = "^5.0.4"
black = "^23.1.0"
ipywidgets = "^7.7"
ipykernel = "^6.21.2"
pyppeteer = "^1.0.2"
jupyterlab = "^3.6.1"
pyparsing = "^3.0.9"
types-jsonschema = "^4.4.1"
pandas-stubs = "^1.2.0"

[tool.isort]
line_length = 88
multi_line_output = 3
include_trailing_comma = true
known_third_party = ["click"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
